Rename Http class to Connection
This commit changes theHttp class to be renamed to Connection; this
fits in better with the `Faraday::Connection` class that it depends
upon. This PR also removes some of the helper methods that were
unnecessary, as they methods simply wrapping settings on the Faraday
object itself.

Also included in this commit is a simple test that demonstrates how
to use the Http class in a test, installing the puppetserver and making
a simple request to the environments endpoint.