Remove argument requirement for Beaker::Host objects
This changes the connection object to not require a Beaker::Host object
for initialization. This change allows for any host to be connected to,
not just SUTs.

This also adds DSL methods for instantiating these objects in the
WebHelpers module. The DSL generation method takes an optional
Beaker::Host object. Also include in the DSL is an `https_request`
method, which is convenient for sending single requestswithout a long
lived object in your test.